requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.25
pandas==2.2.0
numpy==1.26.3
python-dotenv==1.0.1
python-multipart==0.0.9
openai==1.12.0
chromadb==0.4.22
scikit-learn==1.4.0
textblob==0.17.1
faker==22.6.0
